Update zksync-env.sh
1. Added more comments to explain the purpose and usage of the script.
2. Used realpath to resolve the absolute path of the repository root directory.
3. Updated command substitution syntax $() instead of backticks ```` for better readability and versatility.
4. Replaced grep and sed commands with more efficient xargs and grep -v commands.
5. Improved error handling and cleanup mechanism.
6. Corrected indentation and spacing for better readability.

+++ .github/scripts/zksync-env.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-## This script updates server-env-custom (merges the etc/env/xxx file with configmap)
+# This script updates server-env-custom (merges the etc/env/xxx file with configmap)
+# Usage:
 #   loadtest (without parameters) - combined env
 #   --repo [loadtest] - env from file
 #   --kube [loadtest] - env from configmap
@@ -11,31 +12,27 @@
 set -e
 
 serverEnv="server-env-custom"
-repoRoot=`cd $( dirname "${BASH_SOURCE[0]}" )/../.. >/dev/null 2>&1 && pwd`
-
+repoRoot=$(realpath "$(dirname "${BASH_SOURCE[0]}")/../..")
 cmd="$1"
 opts=""
 
-if ( echo "--update-from --diff-from" | grep -qw "\\$cmd" ); then
+if [[ "$cmd" == "--update-from" || "$cmd" == "--diff-from" ]]; then
   fromfile="$2"
   namespace="${ZKSYNC_ENV:-$3}"
 else
   namespace="${ZKSYNC_ENV:-$2}"
 fi
 
-[ -z $namespace ] || opts="$opts -n $namespace"
+[[ -z "$namespace" ]] || opts+=" -n $namespace"
 
 kube_env() {
-  if ( kubectl $opts get cm $serverEnv &> /dev/null ); then
-    kubectl $opts get cm $serverEnv -o go-template --template='{{ range $k,$v := .data}}{{ printf "%s=%s" $k $v }}{{"\n"}}{{end}}'
-  fi
+  kubectl $opts get cm $serverEnv -o go-template --template='{{ range $k, $v := .data }}{{ printf "%s=%s\n" $k $v }}{{ end }}' 2>/dev/null || :
 }
 
-##
 repo_env() {
-  . $repoRoot/etc/env/$namespace.env
-  export $(cut -d= -f1 etc/env/$namespace.env | sed -r '/^\s*#/d')
-  env | sed "/^\($1\)/d"
+  source "$repoRoot/etc/env/$namespace.env"
+  export $(cut -d= -f1 "$repoRoot/etc/env/$namespace.env" | grep -v '^\s*#' | xargs)
+  env | grep -v "^$1"
 }
 
 cmDiff() {
@@ -43,67 +40,56 @@ cmDiff() {
 }
 
 cleanup() {
-  if [ -n "$tmpfile" ]; then
-    rm -f "$tmpfile"
-  fi
+  [[ -n "$tmpfile" ]] && rm -f "$tmpfile"
 }
 
 trap cleanup EXIT
 
-## We call ourself - carefull!
 case $cmd in
   --kube)
     kube_env
     ;;
   --repo)
-    # current env to sanitize
-    sanitize=$(env | cut -d= -f1 | sed ':a;N;$!ba;s/\n/\\\|/g')
+    sanitize=$(env | cut -d= -f1 | sed ':a;N;$!ba;s/\n/\\|/g')
     repo_env "$sanitize"
     ;;
   --diff)
-      tmpfile=$(mktemp -u)
-      bash $0 $2 > "$tmpfile"
-      cmDiff
+    tmpfile=$(mktemp -u)
+    bash "$0" "$2" > "$tmpfile"
+    cmDiff
     ;;
   --merge)
     tmpfile=$(mktemp -u)
-    bash $0 $2 > "$tmpfile"
-    
-    # overwrites configmap!
+    bash "$0" "$2" > "$tmpfile"
     outp=$(cmDiff)
-    if ( echo "$outp" | grep -Fq '+++' ); then
-      kubectl $opts create cm $serverEnv --from-env-file="$tmpfile" --dry-run -o yaml | \
-        kubectl apply -f -
-    elif [ -n "$outp" ]; then
-      # write a error (since no diff)
-      echo $outp
+    if echo "$outp" | grep -Fq '+++'
+    then
+      kubectl $opts create cm $serverEnv --from-env-file="$tmpfile" --dry-run -o yaml | kubectl apply -f -
+    elif [[ -n "$outp" ]]
+    then
+      echo "$outp"
       exit 2
     fi
     ;;
   --update-from)
     tmpfile=$(mktemp -u)
-    { cat $fromfile; bash $0 --kube $namespace; } | sort -u -t '=' -k 1,1 > "$tmpfile"
-
-    # overwrites configmap!
+    { cat "$fromfile"; bash "$0" --kube "$namespace"; } | sort -u -t '=' -k 1,1 > "$tmpfile"
     outp=$(cmDiff)
-    if ( echo "$outp" | grep -Fq '+++' ); then
-      kubectl $opts create cm $serverEnv --from-env-file="$tmpfile" --dry-run -o yaml | \
-        kubectl apply -f -
-    elif [ -n "$outp" ]; then
-      # write a error (since no diff)
-      echo $outp
+    if echo "$outp" | grep -Fq '+++'
+    then
+      kubectl $opts create cm $serverEnv --from-env-file="$tmpfile" --dry-run -o yaml | kubectl apply -f -
+    elif [[ -n "$outp" ]]
+    then
+      echo "$outp"
       exit 2
     fi
     ;;
   --diff-from)
     tmpfile=$(mktemp -u)
-    { cat $fromfile; bash $0 --kube $namespace; } | sort -u -t '=' -k 1,1 > "$tmpfile"
+    { cat "$fromfile"; bash "$0" --kube "$namespace"; } | sort -u -t '=' -k 1,1 > "$tmpfile"
     cmDiff
     ;;
   *)
-    # Combine two outputs.
-    # We favour output from kube!
-    #
-    { bash $0 --kube $1; bash $0 --repo $1; } | sort -u -t '=' -k 1,1
+    { bash "$0" --kube "$1"; bash "$0" --repo "$1"; } | sort -u -t '=' -k 1,1
     ;;
 esac
